requirements.txt
google-genai
streamlit
feedparser
requests
pandas
plotly
python-dotenv